backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
opencv-python-headless
ImageHash
Pillow
numpy
python-dotenv
requests
